+++ python/lsst/ts/ess/controller/device/vcp_ftdi.py
@@ -22,8 +22,8 @@
 __all__ = ["VcpFtdi"]
 
 import asyncio
-import concurrent
 import logging
+import re
 from typing import Callable
 
 from lsst.ts.ess import common
@@ -75,6 +75,13 @@ class VcpFtdi(common.device.BaseDevice):
             auto_detach=False,
         )
 
+        # Build a regular expression to use when checking if the sensor has
+        # sent the end of a telemetry string. Occasionally an additional
+        # NULL character may show up so we need to check for that.
+        enhanced_terminator = ".?".join(self.sensor.terminator)
+        self.enhanced_terminator_regex = re.compile(enhanced_terminator)
+        self.terminator_regex = re.compile("^.*" + enhanced_terminator + "$")
+
     async def basic_open(self) -> None:
         """Open the Sensor Device.
 
@@ -110,9 +117,9 @@ class VcpFtdi(common.device.BaseDevice):
         line: str = ""
         # get running loop to run blocking tasks
         loop = asyncio.get_running_loop()
-        with concurrent.futures.ThreadPoolExecutor(max_workers=1) as pool:
-            while not line.endswith(self.sensor.terminator):
-                line += await loop.run_in_executor(pool, self.vcp.read, 1)
+        while not self.terminator_regex.match(line):
+            line += await loop.run_in_executor(None, self.vcp.read, 1)
+        line = self.enhanced_terminator_regex.sub(self.sensor.terminator, line)
         self.log.debug(f"Returning {self.name} {line=}")
         return line