+++ doc/version_history.rst
@@ -6,6 +6,19 @@
 Version History
 ###############
 
+v0.8.5
+======
+
+* Fix the conda recipe.
+* Remove superfluous use of a ThreadPoolExecutor.
+* Check for a terminator that may contain additional NULL characters.
+
+Requires:
+
+* ts_ess_common
+* ts_tcpip 1.1
+* ts_utils 1.0
+
 v0.8.4
 ======